These files might or might not have to do with the following:
year is wrong

+++ work/src/org/kuali/kfs/gl/batch/SufficientFundsRebuilderStep.java
@@ -37,7 +37,7 @@ public class SufficientFundsRebuilderStep implements Step {
     }
 
     public boolean performStep() {
-        sufficientFundsRebuilderService.rebuildSufficientFunds(new Integer(2006));
+        sufficientFundsRebuilderService.rebuildSufficientFunds(new Integer(2004));
         return true;
     }
